pkg/http: move out http from instrumentation pkg to its own
Move http pkg directory out from instrumentation pkg in order to allow
it to be used in other parts of the nri-resmgr.
